renames translations file
