+++ src/app/components/LiteSiteSummary/index.tsx
@@ -7,7 +7,7 @@ import Text from '../Text';
 import { ServiceContext } from '../../contexts/ServiceContext';
 import { RequestContext } from '../../contexts/RequestContext';
 import styles from './index.styles';
-import defaultTranslations from './liteSiteTranslations';
+import defaultTranslations from './defaultTranslations';
 import CallToActionLink from '../CallToActionLink';
 
 const LiteSiteSummary = () => {
